requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
openai==1.3.7
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1